refactor time filters
