+++ frontend/components/UI/RadioGroup.tsx
@@ -1,12 +1,8 @@
 import { ChangeEvent } from 'react';
-import Button from '@mui/material/Button';
-import Radio from '@mui/material/Radio';
-import RadioGroup from '@mui/material/RadioGroup';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import FormControl from '@mui/material/FormControl';
+import { Button, FormControl, FormControlLabel, Radio, RadioGroup } from '@mui/material';
 import { RadioGroupProps } from '@/assets/interfaces';
 
-export default function RadioButtonsGroup({
+export function RadioButtonsGroup({
   value,
   label,
   setValue,
