+++ frontend/components/UI/index.tsx
@@ -5,3 +5,5 @@ export * from '@/components/UI/DataTable';
 export * from '@/components/UI/EventListing';
 export * from '@/components/UI/Modal';
 export * from '@/components/UI/PopOverButton';
+export * from '@/components/UI/RadioGroup';
+export * from '@/components/UI/SliderDouble';
